fixed #8249 - [Ecommerce] Remove Legacy BatchProcessing workermode
