+++ bundles/EcommerceFrameworkBundle/Command/IndexService/ProcessPreparationQueueCommand.php
@@ -16,7 +16,6 @@ namespace Pimcore\Bundle\EcommerceFrameworkBundle\Command\IndexService;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\IndexService;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\IndexUpdateService;
-use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\Worker\AbstractBatchProcessingWorker;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\Worker\ProductCentricBatchProcessingWorker;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\IndexableInterface;
 use Pimcore\Console\Traits\Parallelization;
@@ -133,7 +132,7 @@ class ProcessPreparationQueueCommand extends AbstractIndexServiceCommand
     /**
      * @param string[] $openTenantList a list of tenants for which the workers should be retrieved
      *
-     * @return AbstractBatchProcessingWorker[]
+     * @return ProductCentricBatchProcessingWorker[]
      */
     private function getTenantWorkers(array $openTenantList): array
     {
