+++ bundles/EcommerceFrameworkBundle/DependencyInjection/Configuration.php
@@ -749,14 +749,6 @@ class Configuration implements ConfigurationInterface
                     ->cannotBeEmpty()
                     ->defaultValue(IndexService::class)
                 ->end()
-                // @TODO Pimcore 10 - remove this
-                ->enumNode('worker_mode')
-                    ->values([ProductCentricBatchProcessingWorker::WORKER_MODE_PRODUCT_CENTRIC, ProductCentricBatchProcessingWorker::WORKER_MODE_LEGACY])
-                    ->cannotBeEmpty()
-                    ->setDeprecated('will be removed in Pimcore 10 as then ' . ProductCentricBatchProcessingWorker::WORKER_MODE_PRODUCT_CENTRIC . ' will be default mode.')
-                    ->info('Worker mode for ' . ProductCentricBatchProcessingWorker::class . ' workers.')
-                    ->defaultValue(ProductCentricBatchProcessingWorker::WORKER_MODE_LEGACY)
-                ->end()
                 ->scalarNode('default_tenant')
                     ->cannotBeEmpty()
                     ->defaultValue('default')
