+++ bundles/EcommerceFrameworkBundle/DependencyInjection/PimcoreEcommerceFrameworkExtension.php
@@ -411,9 +411,6 @@ class PimcoreEcommerceFrameworkExtension extends ConfigurableExtension
 
         $container->setParameter('pimcore_ecommerce.index_service.default_tenant', $config['default_tenant']);
 
-        //@TODO Pimcore 10 - remove this
-        $container->setParameter('pimcore_ecommerce.index_service.worker_mode', $config['worker_mode']);
-
         $getterIds = [];
         $interpreterIds = [];
