+++ bundles/EcommerceFrameworkBundle/IndexService/Worker/OptimizedMysql.php
@@ -39,11 +39,10 @@ class OptimizedMysql extends AbstractMockupCacheWorker implements BatchProcessin
      * @param OptimizedMysqlConfig $tenantConfig
      * @param ConnectionInterface $db
      * @param EventDispatcherInterface $eventDispatcher
-     * @param string|null $workerMode
      */
-    public function __construct(OptimizedMysqlConfig $tenantConfig, ConnectionInterface $db, EventDispatcherInterface $eventDispatcher, string $workerMode = null)
+    public function __construct(OptimizedMysqlConfig $tenantConfig, ConnectionInterface $db, EventDispatcherInterface $eventDispatcher)
     {
-        parent::__construct($tenantConfig, $db, $eventDispatcher, $workerMode);
+        parent::__construct($tenantConfig, $db, $eventDispatcher);
 
         $this->mySqlHelper = new Helper\MySql($tenantConfig, $db);
     }
